requirements.txt
crewai
python-dotenv
fastapi
uvicorn
langchain
langchain-groq
pydantic


